docker/celery/requirements.txt
celery
python-irodsclient
redis
